MmsRequest: Avoid concurrency in automatic DDS switches
back-to-back MMS operations (send/download to self, for example) can
trigger simultaneous attempts at DDS switching, especially because we
signal completion before completing restoration of the "real" selection,
and the modems tend not to like it. So before kicking the completion
intent back to the calling app, switch data back and give it
some time to settle.

Change-Id: I72c196cfebd31111abb867f6b6963c29cfd393e2