Update the logic for Fi user to download the MmsMessage
Previously, if Fi user get WAP_PUSH in subscription1 and it switch to
other subscription, it will fail to download the message.

The new logic is if the subId is not active, try to use subscription in
same group.

Test: Flash the devices, verify Mms works for non-Fi devices and Fi
devices

Bug: 141074219
Merged-in: I0a032dbd62277fb23e2d69b2a986e9d38b7f96e0
Change-Id: I0a032dbd62277fb23e2d69b2a986e9d38b7f96e0