Remove call to setCleartextTrafficPermitted
MmsService.onCreate calls setCleartextTrafficPermitted. This call
duplicates the existing declaration in AndroidManifest.xml to accept
clear text traffic. Remove this duplication.

Note: it is not possible to upgrade to using a NetworkSecurityConfig as
this app is injected in the com.android.phone process and as such needs
to have a matching ManifestConfigSource value (see
NetworkSecurityConfigProvider.handleNewApplication).

Bug: 406067628
Test: atest MmsServiceTests
Test: atest MmsServiceRoboTests
Flag: EXEMPT bugfix
Change-Id: I26afffbd954567de7d1555aa7577107a589a7975